Split interfaces.rs into smaller files
